Refactor tx types into individual files
